+++ shared/teams/new-team/wizard/team-purpose.tsx
@@ -1,0 +1,80 @@
+import * as React from 'react'
+import * as Kb from '../../../common-adapters'
+import * as Container from '../../../util/container'
+import * as Styles from '../../../styles'
+import {ModalTitle} from '../../common'
+
+const TeamPurpose = () => {
+  const dispatch = Container.useDispatch()
+  const nav = Container.useSafeNavigation()
+  const onBack = () => dispatch(nav.safeNavigateUpPayload())
+
+  return (
+    <Kb.Modal
+      onClose={onBack}
+      header={{
+        leftButton: <Kb.Icon type="iconfont-arrow-left" onClick={onBack} />,
+        title: <ModalTitle teamname="" title="New team" />,
+      }}
+      allowOverflow={true}
+    >
+      <Kb.Box2
+        direction="vertical"
+        fullWidth={true}
+        style={styles.body}
+        gap={Styles.isMobile ? 'xsmall' : 'tiny'}
+      >
+        <Kb.Text type="BodySemibold">What do you need a team for?</Kb.Text>
+        <Kb.RichButton
+          description="A small group of people, with no initial need for channels."
+          icon="icon-teams-type-squad-64"
+          onClick={() => undefined} //TODO: Implement this
+          title="Friends, family, or squad"
+        />
+
+        <Kb.RichButton
+          description="With multiple roles and channels."
+          icon="icon-teams-type-business-64"
+          onClick={() => undefined} //TODO: Implement this
+          title="A project, business or organization"
+        />
+
+        <Kb.RichButton
+          description="A forum for people who share an interest or cause."
+          icon="icon-teams-type-community-64"
+          onClick={() => undefined} //TODO: Implement this
+          title="A community"
+        />
+
+        <Kb.RichButton
+          description="Start simple and go from there."
+          icon="icon-teams-type-notsure-64"
+          onClick={() => undefined} //TODO: Implement this
+          title="Other/You're not sure"
+        />
+      </Kb.Box2>
+    </Kb.Modal>
+  )
+}
+
+const styles = Styles.styleSheetCreate(() => ({
+  body: Styles.platformStyles({
+    common: {
+      ...Styles.padding(Styles.globalMargins.small),
+      backgroundColor: Styles.globalColors.blueGrey,
+      borderRadius: 4,
+    },
+    isElectron: {minHeight: 326},
+    isMobile: {...Styles.globalStyles.flexOne},
+  }),
+  container: {
+    padding: Styles.globalMargins.small,
+  },
+  wordBreak: Styles.platformStyles({
+    isElectron: {
+      wordBreak: 'break-all',
+    },
+  }),
+}))
+
+export default TeamPurpose
